day 4 progress

+++ backend/controllers/product.controller.js
@@ -36,7 +36,6 @@ export const putProduct = async (req, res) => {
     if(!mongoose.Types.ObjectId.isValid(id)){
         console.log({id})
         return res.status(404).json({success: false, message: $`Product not found {id}`})
-        
     }
 
     try {
@@ -49,11 +48,16 @@ export const putProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+    
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        console.log({id})
+        return res.status(404).json({success: false, message: $`Product not found {id}`})
+    }
 
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success:true, message: "Product deleted"});
     } catch (error) {
-        res.status(400).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server Error"});
     }
 };
